# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Tesis Guvenlik & Operasyon API
  version: "0.1.0"
  description: |
    Multi-tenant tesis guvenlik & operasyon SaaS — v0 sozlesmesi.
    Backend (FastAPI) ile Mobil (Flutter) ve Panel (Next.js) gelistiricilerinin
    paralel calismasi icin tek dogruluk kaynagidir.

    Konvansiyonlar (detay: /contracts/README.md):
      * Tum zamanlar ISO8601 UTC (`...Z`). tenant.timezone ile yorumlanir.
      * Hata zarfi: `{ "error": { "code": "...", "message": "..." } }`.
      * Sayfalama: `limit` (varsayilan 50, max 200) + `offset`. Liste yanitlari
        `PageMeta` icerir.
      * Auth: JWT Bearer (bkz. /contracts/auth.md). Tenant her zaman token'dan
        cikarilir — istemci tenant_id gondermez/secemez.
      * `POST /scans` icin `Idempotency-Key` header zorunlu.
servers:
  # NOT: base path /v0 KALDIRILDI — gercek backend endpoint'leri /v0 ONEKI OLMADAN
  # sunuyor (orn. /auth/login, /scans). Base = host:port.
  - url: https://api.example.com
    description: production
  - url: http://localhost:8000
    description: yerel gelistirme
...
security:
  - bearerAuth: []
paths:
...
  /assets:
    get:
      tags: [assets]
      summary: Demirbas listesi
      description: |
        Her item ACIK zimmet ozeti tasir: `acik_zimmet` (null | {alan_user_id,
        alan_user_ad, alinma_zamani}) — mobil history taramadan gosterir
        (mobil §13 #2/#5). `nfc_tag_uid` TAM eslesme filtresi UID->asset
        cozumunu tek istege indirir (mobil §13 #1). `checked_out_by` acik
        zimmet filtresi: `me` = token kullanicisinin uzerindekiler (mobil §13
        #3); `<uuid>` degeri YALNIZ admin (baskasi 403), gecersiz deger 422.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: kategori
          in: query
          schema: { $ref: "#/components/schemas/AssetKategori" }
        - name: durum
          in: query
          schema: { $ref: "#/components/schemas/AssetDurum" }
        - name: aktif
          in: query
          schema: { type: boolean }
        - name: nfc_tag_uid
          in: query
          description: Tam eslesme — UID okutup asset'i tek istekle cozmek icin
          schema: { type: string }
        - name: checked_out_by
          in: query
          description: "'me' veya user UUID (UUID yalniz admin) — acik zimmeti bu kullanicida olanlar"
          schema: { type: string }
      responses:
        "200":
          description: Liste
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Asset" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [assets]
      summary: Demirbas olustur
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssetCreate" }
      responses:
        "201":
          description: Olusturuldu
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: nfc_tag_uid bu tenant'ta zaten kayitli
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /assets/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      tags: [assets]
      summary: Demirbas detay
      responses:
        "200":
          description: Detay
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [assets]
      summary: Demirbas guncelle (kismi)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssetUpdate" }
      responses:
        "200":
          description: Guncellendi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      tags: [assets]
      summary: Demirbas sil
      responses:
        "204": { description: Silindi }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /assets/{id}/checkout:
    parameters:
      - $ref: "#/components/parameters/PathId"
    post:
      tags: [assets]
      summary: Demirbasi al (zimmetle)
      description: |
        Idempotency-Key ZORUNLU (scan deseni). nfc_tag_uid verilirse asset'in
        nfc'siyle eslesmeli. alan_user_id token'dan. Zaten zimmetliyse 409.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckoutRequest" }
      responses:
        "201":
          description: Zimmetlendi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetCheckout" }
        "200":
          description: Idempotent tekrar — mevcut kayit
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetCheckout" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Demirbas zaten zimmetli (veya ayni key farkli govde)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /assets/{id}/checkin:
    parameters:
      - $ref: "#/components/parameters/PathId"
    post:
      tags: [assets]
      summary: Demirbasi birak (zimmeti kapat)
      description: |
        Acik zimmeti kapatir (birakma_zamani=now). Idempotency-Key ZORUNLU.
        nfc_tag_uid verilirse asset'in nfc'siyle eslesmeli. Acik zimmet yoksa 409.

        SAHIPLIK (mobil §13 #6): checkin'i YALNIZ acik zimmetin sahibi
        (alan_user_id == token user) veya ADMIN (yonetim mudahalesi)
        yapabilir; baska kullanici 403 `forbidden`
        ("Zimmet baskasinin uzerinde..."). security/tesis_gorevlisi baskasinin
        zimmetini KAPATAMAZ.

        Kapatan kullanici `birakan_user_id` olarak kaydedilir (cevapta ve
        history'de `birakan_user_ad` ile birlikte doner).
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckinRequest" }
      responses:
        "200":
          description: Birakildi (zimmet kapatildi / idempotent tekrar)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetCheckout" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Acik zimmet yok (demirbas zaten musait)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /assets/{id}/history:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      tags: [assets]
      summary: Demirbas zimmet gecmisi
      description: |
        Siralama `alma_zamani` uzerinden; VARSAYILAN `desc` (en yeni ustte —
        mobil §13 #4). Eski (en eski ustte) davranis icin `?order=asc`.
        Item'larda `alan_user_ad` dolu gelir (id + ad birlikte, mobil §13 #5).
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: order
          in: query
          description: "alma_zamani sirasi (varsayilan: desc = en yeni ustte)"
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: Gecmis
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/AssetCheckout" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ================================================================= #
  # EMERGENCY (acil durum) + tenant settings (yonetim numarasi)
  # ================================================================= #
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. `Authorization: Bearer <token>`. Bkz. /contracts/auth.md"
  parameters:
    Limit:
      name: limit
      in: query
      description: Sayfa boyutu (varsayilan 50, max 200)
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      description: Atlanan kayit sayisi
      schema: { type: integer, minimum: 0, default: 0 }
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Istemci tarafindan uretilen benzersiz anahtar (UUID onerilir)
      schema: { type: string, minLength: 8, maxLength: 200 }
  responses:
    BadRequest:
      description: Gecersiz istek
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Kimlik dogrulama gerekli / gecersiz token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Rol bu islem icin yetkili degil
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Kayit bulunamadi (tenant kapsaminda)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Cakisma (benzersizlik / iliskili kayit)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Govde dogrulama hatasi
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Makine-okunabilir hata kodu
              example: validation_error
            message:
              type: string
              description: Insan-okunabilir aciklama
              example: "baslangic_saat bitis_saat'ten once olmali"
            details:
              type: array
              description: Alan bazli dogrulama hatalari (opsiyonel)
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    Page:
      type: object
      required: [meta]
      properties:
        meta: { $ref: "#/components/schemas/PageMeta" }
    PageMeta:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer, example: 50 }
        offset: { type: integer, example: 0 }
        total:
          type: integer
          description: Filtreye uyan toplam kayit sayisi
          example: 137
...
    AssetKategori:
      type: string
      enum: [ekipman, arac, alet, diger]
    AssetDurum:
      type: string
      enum: [musait, zimmetli, bakimda]
    Asset:
      type: object
      required: [id, ad, durum, aktif, created_at]
      properties:
        id: { type: string, format: uuid }
        ad: { type: string, example: "Cim bicme makinesi" }
        kategori: { $ref: "#/components/schemas/AssetKategori" }
        nfc_tag_uid: { type: string, nullable: true }
        durum: { $ref: "#/components/schemas/AssetDurum" }
        aciklama: { type: string, nullable: true }
        aktif: { type: boolean }
        acik_zimmet:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/AcikZimmet" }]
          description: "Acik zimmet ozeti; zimmetli degilse null (mobil §13 #2/#5)"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AcikZimmet:
      type: object
      required: [alan_user_id, alan_user_ad, alinma_zamani]
      properties:
        alan_user_id: { type: string, format: uuid }
        alan_user_ad: { type: string, example: "Guard A" }
        alinma_zamani: { type: string, format: date-time }
    AssetCreate:
      type: object
      required: [ad]
      properties:
        ad: { type: string }
        kategori: { $ref: "#/components/schemas/AssetKategori" }
        nfc_tag_uid: { type: string, nullable: true }
        aciklama: { type: string, nullable: true }
        aktif: { type: boolean, default: true }
    AssetUpdate:
      type: object
      minProperties: 1
      properties:
        ad: { type: string }
        kategori: { $ref: "#/components/schemas/AssetKategori" }
        nfc_tag_uid: { type: string, nullable: true }
        durum: { $ref: "#/components/schemas/AssetDurum" }
        aciklama: { type: string, nullable: true }
        aktif: { type: boolean }
    AssetCheckout:
      type: object
      required: [id, asset_id, alan_user_id, alma_zamani, idempotency_key, created_at]
      properties:
        id: { type: string, format: uuid }
        asset_id: { type: string, format: uuid }
        alan_user_id:
          type: string
          format: uuid
          description: "Token'dan turetilir (alan kullanici)"
        alan_user_ad:
          type: string
          nullable: true
          description: "Alan kullanicinin adi (id + ad birlikte; mobil §13 #5)"
        birakan_user_id:
          type: string
          format: uuid
          nullable: true
          description: "Zimmeti kapatan (sahibi veya admin); acik zimmette NULL"
        birakan_user_ad:
          type: string
          nullable: true
          description: "Birakan kullanicinin adi (id + ad birlikte)"
        alma_zamani: { type: string, format: date-time }
        birakma_zamani:
          type: string
          format: date-time
          nullable: true
          description: "NULL => hala uzerinde (acik zimmet)"
        alma_nfc_tag_uid: { type: string, nullable: true }
        birakma_nfc_tag_uid: { type: string, nullable: true }
        alma_gps_lat: { type: number, format: double, nullable: true }
        alma_gps_lng: { type: number, format: double, nullable: true }
        birakma_gps_lat: { type: number, format: double, nullable: true }
        birakma_gps_lng: { type: number, format: double, nullable: true }
        notlar: { type: string, nullable: true }
        idempotency_key: { type: string }
        created_at: { type: string, format: date-time }
    CheckoutRequest:
      type: object
      properties:
        nfc_tag_uid:
          type: string
          nullable: true
          description: "Okutulan etiket; verilirse asset.nfc_tag_uid ile eslesmeli"
        gps_lat: { type: number, format: double, nullable: true }
        gps_lng: { type: number, format: double, nullable: true }
        notlar: { type: string, nullable: true }
    CheckinRequest:
      type: object
      properties:
        nfc_tag_uid:
          type: string
          nullable: true
          description: "Okutulan etiket; verilirse asset.nfc_tag_uid ile eslesmeli"
        gps_lat: { type: number, format: double, nullable: true }
        gps_lng: { type: number, format: double, nullable: true }
        notlar: { type: string, nullable: true }

    # --------------------------- emergency ------------------------- #
...